Introduce a passpipeline to lower tiled/distributed code through AIR/AIE passes (#26)
This moves the passes used in `aie_flow.sh` to lower the IREE generated tiled code into AIE code. This allows adding a lit test to make sure that the passes work as expected. For now the test starts from an expected input, but this should be generated from a transform dialect script. This change will come as a followup.

To enable this the AIR passes need to be modified to use the auto-generated options object to allows setting the options from C++ as well. This change is currently made in the fork of MLIR-AIR, but is also being upstreamed into the MLIR-AIR repository.

Change the submodule for MLIR-AIE to also point to a fork in nod-ai github. This allows to carry build fix patches for the AIE repo needed by IREE.