[XRT] Add native executable
A flatbuffer format that allows multiple entry points for each
xclbin is added here. Each entry point will have a xrt::kernel
associated with it which will be extracted from the xclbin.
Similarly, each entry point will have a assembly instruction
stream that needs to be passed to the kernel call that is
loaded in the flatbuffer as well.

The parameters required for executable include xrt::kernel and
xrt::bo C++ objects with the asm instructions. Since these cant
be saved in the POD struct we save pointers to these and support
exception handling of this memory allocation process.