Flatbuffer executable serialization (xlcbin, lx6 instructions) (#50)
Write the xclbin and lx6 instructions to a flatbuffer, which eventually
ends up as an .vmfb artefact generated by the compiler. The design for
the AIE backend here is based on the CUDA plugin's design.

This needs testing, once other pieces are connected. I can see the
xclbin and lx6 control code instructions are inserted into the MLIR
module, and the .vmfb file is generated, but have not tested loading and running.
In other words the compilation and artefact generation look fine:

```
./iree-compile -o test.vmfb --mlir-print-ir-after=iree-hal-serialize-executables ... 
```

But to load & run some command queue work to be completed. 
```
./iree-run-module --device=xrt --module=test.vmfb --input=8x16xi32 --input=16x8xi32
```