[amd-aie] add entry_point names to kernels.json and flatbuffer (#58)
Currently the kernel name is hard-coded as "MLIR-AIE". This PR changes
that to take the dispatch entry point function name and also adds that
to the flatbuffer so it can be read at runtime.
I tested this e2e in a sperate branch and will send PRs with the runtime
bits needed to get the matmul working through IREE HAL next.