NFC: General cleanup of IREE-AMD-AIE code and tests.
Most of the passes added recently dont need to run on the
`hal.executable.variant`. They could be run on `func.func`. This
removes a lot of boiler plate from the tests. As a consequence the
workgroup count region handling needs to be moved out of the tile and
fuse pass since it has to be run in a pass rooted on a
`hal.executable.variant`. Move just the workgroup count handling to a
separate pass.

Also organize passes better in header and tablegen files (make them
alphabetical).