[AMD-AIE] Add support for C++ Pack pipeline (#113)
This commit adds the c++ pass pipeline to lower linalg operations using
pack-based approach.
Since we now need to support three separate efforts
now
1) Pad-based lowering
2) Simple Pack-based lowering (now set as default)
3) Pack-based lowering using more advanced optimizations like peeling
and double buffering.

This commit therefore now adds an enum to use to switch between the two pipelines.
The flag `iree-amdaie-use-pipeline` can be used to pick between the three.

Also removed is the use of the CPU pipeline enums from core as a proxy.
Unfortunately, the enums for the pass pipelines setup in core don't make
it extensible in a plugin. This needs to be addressed, but for now, the
"None" value is used, and instead a pack based approach is set as
default. This could be made robust, but will require upstream changes.

---------

Signed-off-by: Abhishek Varma <abhvarma@amd.com>
Co-authored-by: yzhang93 <zhyuhang88@gmail.com>
Co-authored-by: MaheshRavishankar <mahesh@nod-labs.com>