Add a new pack pipeline with loop peeling (#123)
This commit added a pack pipeline that can deal with large k dimension
to fit into the memTile. Note currently the e2e flow (through
MLIR-AIR/MLIR-AIE) doesn't support this pipeline, and the failure is
expected. This pipeline is used to generate the FuncIR which should be
the same output as running the transform dialect.