Ensure cleanup and avoid conflict between jobs (#164)
This gives the deployed xclbins a unique name based on the current
workflow run. This avoids a potential collision where parallel jobs
could overwrite the deployed kernel of another job, thereby un-deploying
it. Then, a cleanup step that runs even if the job is cancelled erases
any kernels that were deployed using the unique name scheme.