Add initial E2E correctness tests (#170)
These tests are compared against a reference matmul implementation with
pseudo random initialization. The test sizes are currently limited. In
upstream IREE similar tests are build using cmake and run using ctest,
however that is currently not possible for a plugin so the script added
here is an alternate way of doing it.