Flags to print IR in mlir-aie passes (i.e. in aie2xclbin) (#191)
Adds 4 flags, which can be used in iree-compile, iree-opt:

```
--aie2xclbin-print-ir-after-all
--aie2xclbin-print-ir-before-all
--aie2xclbin-print-ir-module-scope
--aie2xclbin-disable-threading
```

They behave the same way as the standard MLIR flags (https://mlir.llvm.org/getting_started/Debugging) but only apply to the MLIR pipelines which are run in the external program, aie2xclbin. 

In other words all the IREE and AIR passes are controlled with the usual `--mlir-print-ir-after-all` style flags, but the passes run in mlir-aie are controlled now with `--aie2xclbin-print-ir-after-all style` flags. 

Other flags can be added as and when needed. 

Note that multi-threading is enabled by default in aie2xclbin (just like the default in MLIR).