Vectorization pass, with e2e bf16 test (#172)
This new pass is based on IREE's GenericVectorization pass. This new
pass differs in that

(i) it includes only what is necessary to get a simple matmul example
lowered end-to-end
(ii) it explicitly does not vectorize linalg.copy and linalg.fill (see
code comments for reasons)
(iii) it explicitly does not vectorize operations on 32/64-bit types
(see code comments).


An end-to-end test for bfloat16 is enabled in the matmul test script.
(m=n=64, k=16). Two other tests are (1) with iree-compile, a test to
check for presence of vector.contract and (2) with iree-opt, running
just this new pass to check (ii) and (iii) above.

There is follow-up work to do:

1) check support for integer types (AIE can do "tensor core" matmul for
various int types)
2) check vectorization pipeline for convolution
3) check vectorization where result type is bf16 (current test has
result type of matmul f32)
4) check what the issue is with vectorizing linalg.fill 
5) vectorization of elementwise operations
...