Expand the logic to find tiling and packing config (#225)
This PR adds a largest factor based logic to select tile sizes and
packing config for the pad-pack pipeline. With this arbitrary shapes
seem to be working in local testing. Some CI e2e tests with arbitrary
shapes are added here. Since the other legacy pipelines are not expected
to support such shapes we have to have a new grouping for the shapes.