Option to disable vectorization  (#234)
While it will be unusual to not use the default here (bfloats and small
integer types **must** currently be vectorized in all situations because
AIE does not support scalar operations on these types. And types of size
>= 32 bits **must not** be vectorized because AIE does not have vector
operations for these), it might be useful for debugging purposes to
disable vectorization. If/when peano can emulate scalar bfloat
operations, this might be especially useful to switch between vectorized
and unvectorized lowering.