Deprecate pad and simple-pack pipelines (#239)
We want to deprecate the pad and simple-pack pipelines to lower the
maintenance requirement. From now on, we will be focusing on two
pipelines

- Pad-pack pipeline: is the current default pipeline, which can be
lowered through AIR and run e2e on hardware.
- Pack pipeline: using pack operations for data movement from L3->L2,
L2->L1, and peeling operation to enable initialization in L1. Currently
this pipeline is not fully optimized to run through AIR, and is used for
funcIR/future development.