Add logical objectfifo DMA ops and conversion from AIR (#267)
This is a first PR in a sequence to bring in compilation through logical
objectfifos, using a peeled matmul as an E2E example. See following
branch for all new functionality:
https://github.com/nod-ai/iree-amd-aie/tree/jornt_peeled_matmul_e2e.
This PR adds basic infrastructure for AMDAIE ops with two core ones:
- `AMDAIE::LogicalObjectFifoFromMemrefOp`: a new op which creates a
logical objectfifo from a memref. The logical objectfifo encapsulates
the memref and adds asynchronous access to it through produce/consume or
acquire/release ops, which will be added later on.
- `AMDAIE::DmaCpyNdOp`: Similar to the AIR DMA op this represents a
strided ND copy operation, but operates on logical objectfifos and with
an index result.

To demonstrate usage of these new ops, this PR adds a pass that converts
`AIR::DmaMemcpyNdOp` into `AMDAIE::DmaCpyNdOp`. This is the first
lowering pass run after `pack-to-dma` and `copy-to-dma` in this new
pipeline demonstrated in the branch above.

@nirvedhmeshram @yzhang93 @Abhishek-Varma , could you help with
reviewing?

cc @MaheshRavishankar