Add pass which inserts AIE workgroup and cores (#280)
Adds a pass which inserts AIE workgroup ops around `scf::forall` with
thread mapping (can be extended to blocks as well in the future). The
workgroup consists of two parts:

1. A control code region which will contain the uController control code
instructions to drive the AIE array.
2. An AIE configuration region (I call this an AIE state machine), which
will be converted into AIE cores' elfs + a set of fixed dma
instructions.

The idea of creating this workgroup is that subsequent passes will:
- Lower the AIE configuration region to AIE core and fixed DMA
instructions
- Add required control and synchronization ops into the control code
region
- Move external temporal loops into the control code region
- Recognize and fuse sibling workgroups

This pass comes with some new ops:

- `WorkgroupOp`: designates a region around the parallel loops selected
for parallelization across the AIE array dimensions (columns, rows). For
now, the loops with thread mapping are selected, but this can be
extended in the future to blocks as well.
- `ControlCodeOp`: each workgroup will contain one control code region
op which will contain the uController control code instructions to drive
the AIE array. This operation in inserted in every `WorkgroupOp` through
the build methods to avoid workgroups without corresponding control
code. The `ControlCodeOp` is the terminator as well for easy
retrievability + deterministic location.
- `CoreOp`: Represents an AIE core on a certain location (column, row)
within the AIE array and the operations which should be executed on this
core. Note that this operation differs from the `CoreOp` defined in the
AIE dialect in that this op accepts the column and row as indices
instead of integers (via a TileOp in the AIE dialect). Eventually this
`AMDAIE_CoreOp` will be converted in the `AIE_CoreOp` when we have
static coordinates.
- `EndOp`: A generic terminator for AMDAIE ops' regions, used inside the
`CoreOp`.
- `LogicalObjectFifoConsume`: Operation consuming the result of a DMA
operation. This is a blocking operation,
waiting for the DMA to produce data. Typically, this operation will
reside inside a `CoreOp` to synchronize with external DMA operations
producing data into the respective core's local memory.
- `LogicalObjectFifoProduce`: Operation producing the input of a DMA
operation. This is a release-type operation,
where the DMA will be waiting for the data to be produced. Typically,
this operation will reside inside a `CoreOp` to synchronize with
external DMA operations waiting for data from the respective core's
local memory to be released.