Add workgroup unroll and logical objectfifo distribution pass (#302)
Adds a pass which unrolls the threaded `scf.forall` within a workgroup,
but looks for broadcast capability of dma ops and tries to hoist them
out of the loops if possible. When dma ops are unrolled, new logical
objectfifos are inserted to allow them to be distributed on different
physical tiles.

This pass consists of following steps:

1. Transform threaded `scf.forall` into `scf.for` (this transformation
exists in upstream and will be pulled in from there once changes
propagate)
2. After the `scf.forall` to `scf.for` transformations, typically all
`affine.apply` are located within the inner `scf.for` and this
transformation will try to hoist them to accommodate dma op hoisting.
3. Unroll the `scf.for`'s within a workgroup. Try to hoist dma ops that
don't depend on the loops' induction variables to avoid duplicated dma
ops. Unroll and distribute the remaining logical objectfifos in the
`scf.for` loops.
4. Assign tiles to the logical objectfifos with local memory space
(L1/core).
5. Assign physical AIE tiles to the remaining logical objectfifos. This
rewrite takes an iterative approach by matching logical objectfifos and
only assigning tiles when linked through dma ops with other logical
objectfifos which already have tiles assigned. If the linked logical
objectfifos don't have tiles assigned yet, we will return a failure and
give the linked logical objectfifos a chance to assign tiles before
returning to this one.