Add pass to create a single AIE workgroup (#316)
Adds a pass to create a single AIE workgroup. The workgroup will consist
of:

1. A set of AIE core operations, with dataflow specified by circular DMA
operations.
2. A control code block containing the NPU uController instructions.

This PR adds a few new `amdaie` operations:

- `NpuDmaCpyNdOp`: Represents a strided copy operation with an unlimited
number of dimensions, executed by the Npu uController. This operation
refers to a `CircularDmaCpyNdOp`, which will contain the necessary
information about the source and target logical object fifos of the
operation and which will instantiate the DMA connection.
- `NpuDmaWaitOp`: This wait operation will block on the referenced Npu
DMA operation to complete execution on the provided `direction`. The
`S2MM` direction will block on the destination side of the dma
operation, ensuring complete execution. The `MM2S` direction will block
on the source side of the dma operation, ensuring that the DMA has
successfully started execution, but not guaranteeing that all data has
been received on the destination side.

A few notes:

- I noticed that it's probably not necessary to create/insert workgroups
before this pass, so I will probably get it out of `InsertAIEWorkgroup`
in a follow-up. This will also resolve the unclarity of having both a
`InsertAIEWorkgroup` and `CreateAIEWorkgroup` pass. This inconsistency
is the result of some rewrites/refactoring after getting in the earlier
transformations.
- After this pass, we're very close to MLIR-AIE and except for a few
utility transformations, the next pass will take this code all the way
to MLIR-AIE.