Make loops-for-vectorization recognize matmul_transpose_b (#331)
With the current pipeline, matmul_tranpose_b ops are converted to
matmul-like generics here:
https://github.com/nod-ai/iree-amd-aie/blob/784dedb18b9768ebfff7b344906ee9052e8674eb/compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/AMDAIEPackAndTranspose.cpp#L121
(this is for the pad-pack pipeline, not sure how peel-pack pipeline
differs). As a result, by the time the vectorization pass runs, there
are no matmul_transpose_b operations (or equivalent linalg.generics) in
the IR.

I would to change this (in a future PR). Specifically I would like to
change inner permutation of packing (see
https://github.com/nod-ai/iree-amd-aie/blob/784dedb18b9768ebfff7b344906ee9052e8674eb/compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/KernelDispatch.cpp#L134)
to have no permutation. This will result in matmul_transpose_b not
appearing as matmul in the vectorization passes, but remaining as a
matmul_transpose_b semantically.

The motivation for this (future) change is that performing a permutation
of the inner dimensions doesn't work for data types <32b , because dma
copies cannot have a stride of less than 32 bits. The solution that we
are pursuing is to convert a vector.contract with a transpose map for
the B operand into a vector.transpose operation followed by a
vector.contract with normal matmul maps for operands. Then,
vector.transpose will lower to aievec.shuffle operations (which need to
be implemented, but will supported in hardware (see
[here](https://www.xilinx.com/htmldocs/xilinx2023_2/aiengine_ml_intrinsics/intrinsics/group__intr__gpvectorop__interleave.html)))
and an aievec.matmul operations.