For the pad-pack pipeline, avoid tile size ratio of 2 (#310)
The pad-pack pipeline currently targets a 4x4 AIE array, so for a fixed
shared memory tiling in the M- and N- directions (M0 and N0) it is
better to have

```
M1=M0/4, N1 = N0/4
```
than 
```
M1=M0/2, N1 = N0/2
```

where M1 and N1 are the tiling sizes on the AIE core memory. 

This PR checks for M0/M1 = 2 and if possible halves M1 so that M0/M1 = 4
(ditto the N- dimension)

This PR also reverts the change in #335 to the 'scale factor' chosen in choosing tile sizes. This is required to make the ukernel approach work. We can review this in the future.