Add pass to lower from AMDAIE to AIE/AIEX (#357)
Adds a pass to lower from the AMDAIE dialect to the AIE/AIEX dialects.
This happens in a few steps:

1. Add explicit `amdaie.logicalobjectfifo.link` operations to facilitate
conversion as the AIE dialect relies on this kind of link operations.
2. Create an `AIE::DeviceOp` for every function
3. Convert the HAL binding operations to the NPU instruction function
and ensure the function arguments are in the correct order as indicated
by the HAL binding order (`binding(*)`).
4. Traverse the function and workgroup to build up the `AIE::DeviceOp`.
There are dedicated conversion functions for every AMDAIE operation, but
I would like to highlight two larger ones: `coreToAIE` which traverses
and converts `amdaie.core` and it's body into `aie.core`, and
`controlCodeToAie` which traverses and converts the control code region
and inserts the resulting operations into the NPU instruction sequence
function.
5. Erase all the HAL binding operations and it's uses as they are not
needed after conversion to the NPU function arguments.
6. Visit every `aie.core` operation and ensure it doesn't have any
dependencies outside the core's region by cloning the dependencies and
moving these copies into the core's body. This is typically needed for
constant operations that are used inside the `aie.core` operations.

NOTE: There are quite a few non-direct conversions, i.e. information
from multiple operations or regions is needed to be able to convert into
AIE/AIEX dialect. This happens because the AMDAIE and AIE/AIEX dialects
are set up a bit differently and instead of designing the AMDAIE dialect
to completely map to AIE, I think some changes should be made in the
AIE/AIEX dialects to achieve more natural lowering. See the comments
inside the pass for specific cases. I expect that with some time and
effort on adjusting the dialects on both sides, this lowering can become
more seamless.