Transition AIETarget to use internal aie2xclbin function (#459)
Since https://github.com/nod-ai/iree-amd-aie/pull/420 added a internal
utility to go from AIE to XCLBIN it makes sense for all flows to use
that.
This utility doesn't add any AIE lowering passes so the
`TranslationPassPipeline`s are responsible for that. Hence the passes
needed for the IR lowered through the AIR flow to be further lowered to
AIE are added in its lowering pipeline.

There were some minor additions needed to the utility to get it working
e2e for the AIE flow.

We no longer need to have a separate `aie2xclbin-disable-threading` as
we are passing the existing context and the `mlir-disable-threading`
flag gets automatically propagated in this case. And it actually gives
an error to try changing it if the passed context is multi-threaded.

With this PR the AIE lowering is becoming a part of the IREE
`TranslationPassPipeline` so the pack-peel lit test that was just
passing the air-to-aie conversion had to be watered down to reflect what
can actually be lowered to AIE.

Unfortunately, we still also need the mlir-aie wheel as the internal
utility is still expecting a mlir-aie install for me_basic.o library
[here](https://github.com/nod-ai/iree-amd-aie/blob/main/compiler/plugins/target/AMD-AIE/iree-amd-aie/Target/XCLBinGen.cpp#L302-L304)