[DmaLoopSubsumption] Add option to only support repeat on outer dim (#574)
Fixes an issue exposed by a 64x32x128 matmul:
https://github.com/nod-ai/iree-amd-aie/issues/556.

As the hardware (AIE2(+)) doesn't support `stride == 0` with buffer
descriptors, only a last repeat iteration can be achieved using the
repeat count feature. This PR adjusts the DMA loop subsumption pass to
optionally not perform any subsumption if this would be violated. For
example, the following NPU DMA instruction is not supported on HW:

```
%1 = amdaie.npu.dma_cpy_nd %0([0, 0, %arg3] [8, 2, 16] [16, 0, 1], [] [] [])
```

due to `stride == 0` on the second (middle) dimension. However, the
following is:

```
%1 = amdaie.npu.dma_cpy_nd %0([0, 0, %arg3] [2, 8, 16] [0, 16, 1], [] [] [])
```

but the outer repeat dimension needs to be moved to the first/most outer
dimension for the AIE lowering passes. This is achieved with the logic
added in `LowerToAIE`.