Refactoring the way amdaie_bootgen and amdaie_xclbinutil are found.
Before it was mandatory to specify the --iree-amd-aie-install-dir, because these tools where only being search in specific locations within this folder. This allows to search for the tools in multiple locations, including the env path. Based on the already existing findTool mechanism from iree_compile. It still mantains the old functionality via the flag.