Replace logical objectfifo produce/consume with core in/out DMA operands (#653)
Remove the partially redundant produce/consume ops as they have been
largely replaced already with the `logicalobjectfifo.access` ops. The
main use for them is to provide information on which cores are using
which DMAs which can be handled by adding input and output DMA operands
to the `amdaie.core` ops.