[AMDAIEInsertCore] Handle convolution case  (#658)
Change to get convolution lowering through the insert core pass. Specifically, 
this PR makes sure that linalg::ConvolutionInterfaceOps are handled correctly 
(in the same way as vector::ContractionOp). The pass can be improved further at a
later point (we should be able to lift only a subset of the nested loop
into the core I think).