[Distribute Cores] Do not assume all subviews are 'distributing' (#667)
One role of the distribute-cores-and-objectfifo pass is to private L2
memory allocations.

That is, to replace IR like

```
  %alloc_3 = memref.alloc() : memref<2x2x8x8x4x4xi32, 2 : i32>
  ...
  %subview = memref.subview %alloc_3[%arg2, %arg3, 0, 0, 0, 0] [1, 1, 8, 8, 4, 4] 
                      [1, 1, 1, 1, 1, 1] : memref<2x2x8x8x4x4xi32, 2 : i32> to 
                      memref<1x1x8x8x4x4xi32, strided<[2048, 1024, 128, 16, 4, 1], offset: ?>, 2 : i32>
```

with 

```
  %[[ALLOC_3:.*]] = memref.alloc() : memref<1x1x8x8x4x4xi32, 2 : i32>
```

Before this PR, the pass assumed **all** memrefs on allocs where of this
form. But this is not always the case. This PR tries to determine if the
memref is indeed of this form, otherwise it does nothing.

This is still not completely robust, we either do a proper privatization
pass or get the root cause of the needing this in the first place:
https://github.com/nod-ai/iree-amd-aie/issues/448

This PR also does a refactor of the code.