[LogicalObjectFifoType] Prefer retrieving memory space from type (#675)
Small utility PR to add memory space retrieval to
`LogicalObjectFifoType` and prefers using it as it is more general as it
would work with any value of this type. Needed for a follow-up in which
a new `LogicalObjectFifoPlaceholderOp` will be added, making it
unnecessary to distinguish between this new op and
`LogicalObjectFifoFromMemrefOp` in a lot of cases.