[StatefulTransform] Simplify and refactor around link ops (#754)
Refactors the `AMDAIEStatefulTransform` pass to simplify logic by making
link ops more central. This rewrite makes it easier to start updating it
to work on amdaie dialect ops. For example, the logic on
`aie.objectfifo.link` + `aie.objectfifo` can be replaced by logic on
`amdaie.logicalobjectfifo.from_memref` + `amdaie.connection` ops.

Some notes on removed logic that I think is not needed/exotic for the
foreseeable future:
- Removed support for different buffer depths in single objectFifo
  - Deleted link_test_AIE2
- Removed support for exotic objectFifo consumption/subview access in
cores, for example fifo has depth 4 and is accessed from the core with
depths [3, 2] (4 != 3 + 2). I don't see a use case for this and I am
also suspicious of the correct behaviour of the existing implementation
as the locks don't look to be acquired and released in a
correct/reasonable way in tests.
- Updated subview_test_1 to depth 4 with core AcquireGreaterEqual access
depths [2, 2]
  - Deleted subview_test_2
  - Deleted subview_test_3
- Removed support for reading from shared core tile memory. We can
easily add this back in the future when needed, but it will be easier to
do this after all logic has moved to be performed on amdaie
logicalObjectFifo ops.
- Updated some tests to have additional local buffers where previously a
single buffer was shared between cores
  - Deleted cyclostatic_AIE2_sharedMem