Add `amdaie.flow` for packet flow lowering (#815)
This PR introduces the `amdaie.flow` operation (in consistency with
`aie.flow`). This operation represents connections between
ports/channels on tiles in the array and is lowered to `aie.flow` for
routing. This idea is to lower `amdaie.connection` to `amdaie.flow` +
DMA operations, however this PR goes halfway by creating `amdaie.flow`
operations, while still keeping a reference inside `amdaie.connection`
as the latter is still needed inside `LowerToAIE` for lowering to the
AIE dialect DMA operations. This will be cleaned up in the future, but
for now, this enables me to start adding routes for control packets in
conjunction with data routes for
https://github.com/nod-ai/iree-amd-aie/issues/714.