[ConvertToDma] Add options to tranpose dma dimensions on target side (#812)
Pack/unpack ops change the data layout and thus after converting to dma
ops, the dma addressing dimensions are expanded/collapsed and
transposed. Previously, all the dimension transpositions are on the
source side of dma ops. This PR extends the usage to have an option for
transposition happen on the target side.

In applications, we could make choices of transposition on source or
target for pack or unpack ops based on performance and hardware dma
requirements, etc. The motivation comes from [this
discussion](https://github.com/nod-ai/iree-amd-aie/issues/764#issuecomment-2349286027),
and this PR moves the dma optimization logic to an early pass where the
dma ops are converted.

Note the default options are not changed in this PR (will enable it in a
separate PR with other changes for dma optimization), but I have tested
all four combinations locally to make sure the dma generations are
correct and work e2e. The change of options can be added for example as

```
AMDAIEConvertToDmaOptions dmaOptions;
dmaOptions.packTransposeOnSource = false;
dmaOptions.unpackTransposeOnSource = true;
passManager.addPass(createAMDAIEConvertToDmaPass(dmaOptions));
```