[DistributeCoresAndObjectFifos] Fix for case of mixed operands (subviews and allocs).  (#811)
Fix for a regression in depthwise convolution with packing, when we use
the upstream pass to remove unit extent dimensions
`useRankReducingSlices = true`. Enabling the unit extent dimension
removal results in a op:


```
linalg.generic {indexing_maps = [#map, #map], iterator_types = ["parallel", "parallel"]} ins(%alloc : memref<4x4xi32, 2 : i32>) outs(%subview : memref<4x4xi32, strided<[4, 1]>, 2 : i32>) { ... }
```

appearing in the IR, which is new because it's the first time we see an
op with one operand directly from a `memref.alloc`, and another from a
`memref.subview` operation. This PR adds a fix for this case (by
simplifying the logic...). Before this PR, the assumption was that
all/none of the operands were from `memref.alloc` (I think).

---------

Co-authored-by: Abhishek Varma <abhvarma@amd.com>