Move most of LLVM lowering out of aie2xclbin (#838)
Is a continuation of https://github.com/nod-ai/iree-amd-aie/commit/be42d1918f43ed04595a1384cd9274305e42f961
-- it that PR we moved vector->aievec into the main pipeline, this PR
moves aievec->llvm into the main pipeline (and a few other dialect
lowerings to llvm). Moving the lowering of func to llvm is tricky,
because of the position of the pass `AMDAIECoreToStandardFunc` which
lowers from CoreOp to func. TBC.